complete category pull-down menu to transaction/add page
